Add note about finding process listening on port
I need this infrequently enough that I always forget how to do it.
The easy thing to do is to grep netstat, but I don't even have netstat
installed, and I'd like to keep it that way, because fuser is much more
helpful (it also works for files, and the output is less cryptic). But
I still forget the name of the command, and the CLI syntax it expects.