Suggest gs for pdf to png conversion
Imagemagick has recently been plagued by a policy that disallows it from
processing pdf files by default, which makes sense on servers that do
automated processing of untrusted user input. But I want to convert
files that I produced myself, and "convert" is not doing the trick any
more. But Ghostscript is nice. It is one of those things from the 1980s
that has stood the test of time, and for sure it doesn't print emoji to
my terminal.