Add Unicode chars to copy-paste
Colemak is great, but on Linux the dead key for ... doesn't work. Or
maybe it does, but I don't have it configured, and I keep on opening
unicodeit.net many times per day to get the ellipsis. Better to copy
it from some doc that I can just have open in a window on the side.